requirements.txt
streamlit
pymongo
python-dotenv
pandas
numpy
plotly
scipy>=1.14.1
scikit-learn